pyproject.toml
[tool.poetry]
name = "xero-x-sniper"
version = "0.1.0"
description = "x mention sniper"
authors = ["inakode"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"  # ✅ Fix Python compatibility range
solders = "^0.26.0"
jsonalias = {version = "0.1.1", python = ">=3.7,<4.0"}  # ✅ Explicit Python version marker
jupiter-python-sdk = "^0.0.2.0"
base58 = "^2.1.1"
requests = "^2.28.1"
asyncio = "^3.4.3"
telegram = "^0.0.1"
python-dotenv = "^1.0.0"
tweepy = "^4.15.0"
python-telegram-bot = "^22.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
